refactor: modularize path and linking problem providers
