+++ src/problems/path.ts
@@ -1,0 +1,116 @@
+import { constants, stat } from "fs/promises";
+
+import type { Problem } from "../problems.js";
+import { getRuntimeConfig } from "../runtimeConfig.js";
+import { verifyDir } from "../utils.js";
+
+export type PathIssue =
+        | "missing"
+        | "unreadable"
+        | "unwritable"
+        | "cross-platform-linking"
+        | "linking-failed";
+
+export type PathProblemCategory =
+        | "torrentDir"
+        | "outputDir"
+        | "injectDir"
+        | "linkDirs"
+        | "dataDirs";
+
+export interface PathProblemDescriptor {
+        category: PathProblemCategory;
+        name: string;
+        path: string;
+        issue: PathIssue;
+        message: string;
+        severity: Problem["severity"];
+}
+
+export function buildPathProblem(descriptor: PathProblemDescriptor): Problem {
+        const { category, name, path, issue, message, severity } = descriptor;
+        return {
+                id: `path:${category}:${issue}:${path}`,
+                severity,
+                summary: `${name}: ${message}`,
+                details: `Path: ${path}`,
+                metadata: { category, path, issue },
+        };
+}
+
+export async function diagnoseDirForProblems(
+        path: string,
+        name: string,
+        category: PathProblemCategory,
+        options: { read?: boolean; write?: boolean },
+): Promise<PathProblemDescriptor | null> {
+        const mode =
+                (options.read ? constants.R_OK : 0) |
+                (options.write ? constants.W_OK : 0);
+        const exists = await verifyDir(path, name, mode);
+
+        if (exists) return null;
+
+        try {
+                await stat(path);
+        } catch (error) {
+                if ((error as NodeJS.ErrnoException).code === "ENOENT") {
+                        return {
+                                category,
+                                name,
+                                path,
+                                issue: "missing",
+                                message: "Directory does not exist.",
+                                severity: "error",
+                        };
+                }
+        }
+
+        return {
+                category,
+                name,
+                path,
+                issue: options.write ? "unwritable" : "unreadable",
+                message: options.write
+                        ? "cross-seed cannot write to this directory."
+                        : "cross-seed cannot read from this directory.",
+                severity: "error",
+        };
+}
+
+export async function collectPathProblems(): Promise<Problem[]> {
+        const { torrentDir, outputDir, injectDir } = getRuntimeConfig();
+        const descriptors: PathProblemDescriptor[] = [];
+
+        if (torrentDir) {
+                const descriptor = await diagnoseDirForProblems(
+                        torrentDir,
+                        "torrentDir",
+                        "torrentDir",
+                        { read: true },
+                );
+                if (descriptor) descriptors.push(descriptor);
+        }
+
+        if (outputDir) {
+                const descriptor = await diagnoseDirForProblems(
+                        outputDir,
+                        "outputDir",
+                        "outputDir",
+                        { read: true, write: true },
+                );
+                if (descriptor) descriptors.push(descriptor);
+        }
+
+        if (injectDir) {
+                const descriptor = await diagnoseDirForProblems(
+                        injectDir,
+                        "injectDir",
+                        "injectDir",
+                        { read: true, write: true },
+                );
+                if (descriptor) descriptors.push(descriptor);
+        }
+
+        return descriptors.map(buildPathProblem);
+}
